Restructure utils classes
