+++ pkg/controller/modelutils/k8s/container.go
@@ -9,11 +9,12 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-
 package utils
 
 import (
 	corev1 "k8s.io/api/core/v1"
+	"regexp"
+	"strings"
 )
 
 func BuildContainerPorts(exposedPorts []int, protocol corev1.Protocol) []corev1.ContainerPort {
@@ -28,4 +29,12 @@ func BuildContainerPorts(exposedPorts []int, protocol corev1.Protocol) []corev1.
 		return nil
 	}
 	return containerPorts
+}
+
+var imageRegexp = regexp.MustCompile(`[^0-9a-zA-Z]`)
+
+func GetContainerNameFromImage(image string) string {
+	parts := strings.Split(image, "/")
+	imageName := parts[len(parts)-1]
+	return imageRegexp.ReplaceAllString(imageName, "-")
 }
