+++ pkg/controller/modelutils/k8s/service.go
@@ -9,7 +9,6 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-
 package utils
 
 import (
